refactor CliDrawer and ServerlessProcess into seporate files
